Fix guild-war score GC-packet width on LP64 (sizeof(long) serialization)
The guild-war score is an int at the source (CGuild::GetWarScoreAgainstTo
returns int; STeamData::iScore and the SetWarScoreAgainstTo parameter are
int), but three GC_GUILD/GUILD_SUBHEADER_GC_WAR_SCORE builders serialized
it with sizeof(long) and advertised the sub-packet size with sizeof(long):

  - guild_war.cpp CGuild::SendEnemyGuild  (size :77, long lScore :92, two
    Packet(&lScore, sizeof(long)) :99/:106) -- on the enter-game path
    (Entergame -> LoginMember -> SendEnemyGuild) for members of a guild
    with an active war.
  - guild_war.cpp CGuild::SetWarScoreAgainstTo field-war branch (size :204,
    write(&iScore, sizeof(long)) :211).
  - war_map.cpp  CWarMap::SendScorePacket  (size :647,
    write(&iScore, sizeof(long)) :653).

On FreeBSD LP64 sizeof(long)==8, so the server emitted an 8-byte score
(and a 4-byte-too-large packet size) where the Windows LLP64 client's
4-byte protocol expects 4 -- a per-field stream shift. The two write()
sites also over-read 4 bytes past the 4-byte int source. Serialize as
int32_t (size and width) so the FreeBSD server emits the same 12-byte
body the x86 build and the client protocol expect. Game core rebuilt on
FreeBSD amd64.

Note: the Windows client's CPythonNetworkStream::RecvGuild has no
GUILD_SUBHEADER_GC_WAR_SCORE case and no default, so it never consumes
this sub-packet's body on either platform -- a pre-existing (non-x64)
client desync when an active war exists at login. Tracked separately.

+++ Lead-Server-Source/game/src/war_map.cpp
@@ -644,13 +644,13 @@ void CWarMap::SendScorePacket(BYTE bIdx, LPDESC d)
 
 	p.header = HEADER_GC_GUILD;
 	p.subheader = GUILD_SUBHEADER_GC_WAR_SCORE;
-	p.size = sizeof(p) + sizeof(DWORD) + sizeof(DWORD) + sizeof(long);
+	p.size = sizeof(p) + sizeof(DWORD) + sizeof(DWORD) + sizeof(int32_t);
 
 	TEMP_BUFFER buf;
 	buf.write(&p, sizeof(p));
 	buf.write(&m_TeamData[bIdx].dwID, sizeof(DWORD));
 	buf.write(&m_TeamData[bIdx ? 0 : 1].dwID, sizeof(DWORD));
-	buf.write(&m_TeamData[bIdx].iScore, sizeof(long));
+	buf.write(&m_TeamData[bIdx].iScore, sizeof(int32_t));
 
 	if (d)
 		d->Packet(buf.read_peek(), buf.size());
